requirements.txt
fastapi[standard]
uvicorn
pydantic
sqlalchemy
PyJWT
psycopg2-binary